Many updates
I got the image to show but it would only show 1 emotion 'sad' thought the AI already knew me better than myself. But no it wasnt clairvoyant it was just  because The training model was not fully completing its epochs only 3% through. I managed to fix the model and it does train and I'm expecting 75% accuracy but it takes like 2 hours.